Fixed typo in README.md (#172)
Hi there, 
I noticed the file referenced in the readme doesn't exist under the name specified. 
It looks like an obvious typo so I just decided to correct it. 
Additionally, it seems as if `Alerting Dashboard` shouldn't be there so feel free to make a suggestion on this PR, and then I'll take it out. 
Thanks for this project, really helpful. 
Cheers